fix: torch dtype for Inception model

+++ ltn_model.py
@@ -72,7 +72,7 @@ class CustomDDPM(L.LightningModule):
     
     def validation_step(self, batch):
         real_image, categorical_conds, continuous_conds = self.unfold_batch(batch)
-        real_image = real_image.to(dtype=torch.uint8, device=self.device)
+        real_image = real_image.to(dtype=torch.float32, device=self.device)
         fake_image = self(categorical_conds, continuous_conds, to_save_fig=False)
         
         fake_image = torch.stack([
@@ -85,7 +85,7 @@ class CustomDDPM(L.LightningModule):
                 )
             ).permute(2, 0, 1)
             for f_img in fake_image
-        ]).to(dtype=torch.uint8, device=self.device)
+        ]).to(dtype=torch.float32, device=self.device)
         
         print(f"... {fake_image.dtype = } / {real_image.dtype = }")
         fid = get_fid(fake_image, real_image)
